requirements.txt
fastapi
uvicorn
pytesseract
pillow
streamlit
requests
